move helper function from metaclass to convolve.py
