+++ pyrenew/latent/hospitaladmissions.py
@@ -9,12 +9,9 @@ import jax.numpy as jnp
 import numpyro
 
 import pyrenew.arrayutils as au
+from pyrenew.convolve import compute_delay_ascertained_incidence
 from pyrenew.deterministic import DeterministicVariable
-from pyrenew.metaclass import (
-    RandomVariable,
-    SampledValue,
-    compute_delay_ascertained_incidence,
-)
+from pyrenew.metaclass import RandomVariable, SampledValue
 
 
 class HospitalAdmissionsSample(NamedTuple):
